+++ src/main/java/com/sprint/mission/discodeit/controller/MessageController.java
@@ -7,8 +7,8 @@ import com.sprint.mission.discodeit.dto.response.MessageResponse;
 import com.sprint.mission.discodeit.dto.response.PageResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.service.BinaryContentService;
 import com.sprint.mission.discodeit.service.MessageService;
-import com.sprint.mission.discodeit.service.basic.BinaryContentService;
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
